+++ components/ToastMessage.tsx
@@ -1,0 +1,51 @@
+import React, { useState, useEffect } from 'react';
+import { Text, StyleSheet, Animated } from 'react-native';
+
+const ToastMessage = ({ message, type, visible, onHide }: { message: string; type: 'success' | 'error'; visible: boolean; onHide: () => void }) => {
+  const [opacity] = useState(new Animated.Value(0));
+
+  useEffect(() => {
+    if (visible) {
+      Animated.timing(opacity, {
+        toValue: 1,
+        duration: 300,
+        useNativeDriver: true,
+      }).start(() => {
+        setTimeout(() => {
+          Animated.timing(opacity, {
+            toValue: 0,
+            duration: 300,
+            useNativeDriver: true,
+          }).start(onHide);
+        }, 3000);
+      });
+    }
+  }, [visible]);
+
+  if (!visible) return null;
+
+  return (
+    <Animated.View style={[styles.toast, { opacity, backgroundColor: type === 'success' ? '#4CAF50' : '#F44336' }]}>
+      <Text style={styles.toastText}>{message}</Text>
+    </Animated.View>
+  );
+};
+
+const styles = StyleSheet.create({
+  toast: {
+    position: 'absolute',
+    top: 10,
+    left: 10,
+    right: 10,
+    padding: 10,
+    borderRadius: 5,
+    zIndex: 1000,
+  },
+  toastText: {
+    color: '#FFFFFF',
+    textAlign: 'center',
+    fontWeight: 'bold',
+  },
+});
+
+export default ToastMessage;
